requirements.txt
asgiref==3.9.2
attrs==25.3.0
bleach==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.7
django-ckeditor==6.7.3
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-decouple==2.1
django-extensions==4.1
django-filter==25.2
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modeltranslation==0.19.17
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.11.0
